+++ CHANGELOG.md
@@ -15,7 +15,7 @@ https://tapis-project.github.io/live-docs
 Incremental improvements and new preview features.
 
 ### New features:
-- Support having a dynamic *rootDir* property. Allow *apiUserId* and *effectiveUserId*. Attribute *rootDir* may start with *HOST_EVAL()*.
+- Support having a dynamic *rootDir* property. Attribute *rootDir* may start with *HOST_EVAL()* and may contain *${effectiveUserId}*.
 - For a dynamic *rootDir*, create the resolved *rootDir* on the system host as needed during the credential registration process.
 - Add support for getSystem *resolve* query parameter. Options are *ALL*, *NONE*, *ROOT_DIR*, *EFFECTIVE_USER*.
 - Add attributes returned for getSystem: *isPublic*, *isDynamicEffectiveUser*, *isDynamicRootDir*.